requirements.txt
numpy>=1.24
opencv-python>=4.8
pydantic>=2.6
PyYAML>=6.0
matplotlib>=3.7
pyserial>=3.5
